rover/ntrip-client/requirements.txt
pygnssutils
pyrtcm
pyubx2
pyserial
numpy
influxdb3-python
dotenv